# code/sonara_back_end/swagger/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Sonara API",
    "description": "\nAPI do projeto **Sonara** — plataforma de conexão entre artistas e organizadores de eventos.\n\n## Autenticação\nAlgumas rotas exigem **Bearer Token** (JWT). Faça login em `POST /v1/sonara/usuario/login` e use o token retornado no header:\n`Authorization: Bearer <token>`\n\n## Fluxo principal\n1. Cadastre um usuário (artista ou organizador)\n2. Faça login para obter o token\n3. Crie eventos (organizador) ou candidate-se a eventos (artista)\n4. Gerencie candidaturas, contra-propostas e convites\n        ",
    "version": "1.0.10",
    "contact": {
      "name": "Davi de Almeida Santos"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "Insira: Bearer <token>"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/sonara/usuario/{id}/foto": {
      "patch": {
        "tags": [
          "Usuário"
        ],
        "summary": "Atualizar foto do usuário",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do usuário"
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Nova foto (JPG, PNG ou WEBP)"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "senha": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Foto atualizada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Nenhuma imagem enviada"
          },
          "404": {
            "description": "Usuário não encontrado"
          },
          "415": {
            "description": "Unsupported Media Type"
          }
        }
      }
    },
...
    "/v1/sonara/status/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/v1/sonara/status/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/v1/sonara/foto/": {
      "get": {
        "tags": [
          "Foto"
        ],
        "summary": "Listar todas as fotos",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Foto"
        ],
        "summary": "Upload de foto para evento",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Imagem (JPG, PNG ou JPEG, máx 5MB)"
          },
          {
            "name": "evento_id",
            "in": "formData",
            "type": "integer",
            "required": true,
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Foto enviada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Arquivo ou evento_id ausente/inválido"
          },
          "502": {
            "description": "Erro ao enviar para o Azure"
          }
        }
      }
    },
    "/v1/sonara/foto/{id}": {
      "get": {
        "tags": [
          "Foto"
        ],
        "summary": "Buscar foto por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Foto encontrada"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Foto não encontrada"
          }
        }
      },
      "put": {
        "tags": [
          "Foto"
        ],
        "summary": "Atualizar foto de evento",
        "description": "",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "foto",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "Nova imagem"
          },
          {
            "name": "evento_id",
            "in": "formData",
            "type": "integer",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Foto atualizada"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Foto não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Foto"
        ],
        "summary": "Excluir foto",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Foto excluída"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Foto não encontrada"
          }
        }
      }
    },
    "/v1/sonara/evento/": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Listar todos os eventos",
        "description": "Retorna todos os eventos cadastrados com fotos, endereço e organizador.",
        "responses": {
          "200": {
            "description": "Lista de eventos retornada com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Nenhum evento encontrado"
          }
        }
      },
      "post": {
        "tags": [
          "Evento"
        ],
        "summary": "Criar evento",
        "description": "Cria evento com endereço e vínculo ao organizador.",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Evento criado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Campos obrigatórios ausentes ou inválidos"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/evento/{id}": {
      "get": {
        "tags": [
          "Evento"
        ],
        "summary": "Buscar evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento"
        ],
        "summary": "Atualizar evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoCadastro"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Evento atualizado com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento"
        ],
        "summary": "Excluir evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID do evento"
          }
        ],
        "responses": {
          "200": {
            "description": "Evento excluído com sucesso"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Evento não encontrado"
          }
        }
      }
    },
...
    "/v1/sonara/eventoStatus/": {
      "get": {
        "tags": [
          "Evento Status"
        ],
        "summary": "Listar histórico de status dos eventos",
        "description": "",
        "responses": {
          "200": {
            "description": "Lista retornada"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "post": {
        "tags": [
          "Evento Status"
        ],
        "summary": "Registrar status de evento",
        "description": "",
        "parameters": [
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Status registrado"
          },
          "204": {
            "description": "No Content"
          },
          "415": {
            "description": "Content-Type deve ser application/json"
          }
        }
      }
    },
    "/v1/sonara/eventoStatus/{id}": {
      "get": {
        "tags": [
          "Evento Status"
        ],
        "summary": "Buscar status de evento por ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Status encontrado"
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Evento Status"
        ],
        "summary": "Atualizar status de evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "content-type",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EventoStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Atualizado"
          },
          "204": {
            "description": "No Content"
          }
        }
      },
      "delete": {
        "tags": [
          "Evento Status"
        ],
        "summary": "Excluir status de evento",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Excluído"
          },
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EventoCadastro": {
      "type": "object",
      "properties": {
        "nome": {
          "type": "string",
          "example": "Festival Sonara 2026"
        },
        "descricao": {
          "type": "string",
          "example": "Festival de música ao ar livre"
        },
        "local": {
          "type": "string",
          "example": "Parque Ibirapuera"
        },
        "data": {
          "type": "string",
          "example": "2026-08-15"
        },
        "hora_inicio": {
          "type": "string",
          "example": "18:00"
        },
        "hora_fim": {
          "type": "string",
          "example": "23:00"
        },
        "organizador_id": {
          "type": "number",
          "example": 1
        },
        "cep": {
          "type": "string",
          "example": "04094-050"
        },
        "cidade": {
          "type": "string",
          "example": "São Paulo"
        },
        "estado": {
          "type": "string",
          "example": "SP"
        },
        "logradouro": {
          "type": "string",
          "example": "Av. Pedro Álvares Cabral"
        },
        "numero": {
          "type": "string",
          "example": "S/N"
        },
        "bairro": {
          "type": "string",
          "example": "Vila Mariana"
        },
        "complemento": {
          "type": "string",
          "example": "Portão 3"
        }
      },
      "required": [
        "nome",
        "descricao",
        "local",
        "data",
        "hora_inicio",
        "hora_fim",
        "organizador_id",
        "cep",
        "cidade",
        "estado",
        "logradouro",
        "numero",
        "bairro"
      ]
    },
...
    "EventoStatus": {
      "type": "object",
      "properties": {
        "evento_id": {
          "type": "number",
          "example": 1
        },
        "status_id": {
          "type": "number",
          "example": 1
        },
        "data_hora": {
          "type": "string",
          "example": "2026-06-01 18:00:00"
        }
      },
      "required": [
        "evento_id",
        "status_id",
        "data_hora"
      ]
    },
...
  }
}